Demo uses LSM instead of SM base class + workaround to having a static member function
