+++ lib/local_subnet_manager.hpp
@@ -7,12 +7,30 @@
 #include "component_list.hpp"
 #include "subnet_manager.hpp"
 #include <cstdint>
+#include <functional>
+#include <memory>
+#include <iostream>
 
-class LocalSubnetManager : public SubnetManager
+
+class LocalSubnetManager;
+
+void LSM_messageCallback(std::shared_ptr<LocalSubnetManager> lsm, cubiumServerSocket_t * sock);
+
+class LocalSubnetManager : public SubnetManager, public std::enable_shared_from_this<LocalSubnetManager>
 {
 public:
   LocalSubnetManager(std::shared_ptr<SpaCommunicator> c, LogicalAddress log, uint16_t port) : SubnetManager(c, log, port) {}
-  void receiveMessage(SpaMessage* message);
+  void listenMessages()
+  {
+    if (communicator)
+    {
+      auto callback = std::bind(LSM_messageCallback, LocalSubnetManager::shared_from_this(), std::placeholders::_1);
+      communicator->listen(callback);
+    }
+  }
+  void receiveMessage(SpaMessage* message) {} // Note: Deprecated?
   ComponentList components;
 };
+
+
 #endif
